+++ Sources/ScreenshotScribbler/Layout/LayoutConfig.swift
@@ -14,6 +14,12 @@ public struct LayoutConfig {
     /// Color which covers the whole background. (Default: white)
     public var backgroundColor: ColorType = .solid(color: DefaultColor.CSS.white)
     
+    /// Scaling to apply to the background image, if image is defined. (Default: stretchFill)
+    public var backgroundImageScaling: ImageScaling = .mode(.stretchFill)
+    
+    /// Alignment of the background image, if image is defined. (Default: center, middle)
+    public var backgroundImageAlignment: Alignment = Alignment(horizontal: .center, vertical: .middle)
+    
     /// Percentage of the caption area height relative to the total height of the output image. (Default: 0.25 (25%))
     public var captionSizeFactor: CGFloat = 0.25